requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
exceptiongroup==1.2.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
lxml==5.2.2
multidict==6.0.5
outcome==1.3.0.post0
peewee==3.17.5
psycopg2==2.9.9
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
transliterate==1.10.2
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.1
wsproto==1.2.0
yarl==1.9.4
